api/tesseract/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
python-multipart==0.0.20
pytesseract==0.3.13
pillow==12.0.0
